requirements.txt
annotated-types==0.5.0
anyio==3.7.1
exceptiongroup==1.1.2
fastapi==0.100.0
idna==3.4
mysql-connector-python==8.1.0
protobuf==4.21.12
pydantic==2.0.3
pydantic_core==2.3.0
PyMySQL==1.1.0
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
